WIP: Use the bound of the TypeVariable as search class
Instead of using the TypeVariable itself, we need to use the resolved bound of it. The TypeVariable itself may not be unbounded, but we may already have found bounds for it.
Moreover, if arg or its direct supertypes are not parameterized (Types.param() will return null) we can return false here.