Create isSafeAssignable to help ops.run
We also add a boolean to control isSafeAssignable usage.
We really only want to call isSafeAssignable when we are using
MatchingUtils through the OpMatcher, so there is now a boolean allowing
the user to control whether or not they want to go that extra distance
and call isSafeAssignable or if they just want to check if there will be
no compiler error.