Add collections of functional Op interfaces
These can be used to easily determine the arity of a known functional
interface and allow to match interfaces of different types via these
arities (e.g., Function3 <--> 3 <--> Computer3).
This will be used in the next commit to merge Op transformers that only
operate on interfaces of specific arities into more general
transformers.