Remove Op alias notion
This also removes the notion of a "canonical" Op name. From now on, all
names that are specified for any given Op are treated equally.

Note that this commit also fixes #20. It is however questionable
whether the fixed behavior is desired at all. Might be adjusted in a
future commit.