WIP: Reconcile with underlying framework changes
The ItemIO attribute changed from type to itemIO.
The new framework exposes some errors in ItemIO typing.
There are still some remaining issues, but it is close.

This will probably need to be squashed, along with all commits
of this branch starting at the TEMP imagej-ops dependency commit,
into one single monster patch. We will then apply this patch to
the imagej-ops codebase itself, adding scijava-ops 0.1.0 as a
dependency there.