OpWrappers: retain requested type
When we return an Op from the matcher, it is being assigned to a
particular Op Type (e.g. BiFunction). We need to remember this type
(BiFunction) and not whatever type the Op actually is (i.e. not the
    net.imagej.ops.math.add.BiFunctionAdder type). This will allow the
returned op to be passed back into the matcher with no error