Adapt Ops replacing transformations: first cut
We also generate a bunch of elementary adapt ops

Fixes made to the matcher in the process:

* Improve any checking in parameterizedTypes

Now, a Function<Function<I, O>, Function<Iterable<I>, Iterable<O>>> can
be matched to a Function<Any, Function<Iterable<Double>,
   Iterable<Double>>>

* Improve TypeVariable assignability

Sometimes TypeVariables are assigned to Anys and sometimes this is
correct. Sometimes, however, this is incorrect, and we can later assign
this TypeVariable to a concrete object. (We do have to be careful,
    though, to make sure that the TypeVariable actually is assignable to
    that Object).