Use supertype to infer type variables
Sometimes, when inferring the type variables, we try to do so using some
supertype, and that supertype may have more type arguments that the type
we need to infer. Think of imglib2's RAI<I> and
ArrayImg<FloatType, FloatArray>. Attempting to infer the I would result
in a TypeInferenceException because the ArrayImg has two Type arguments
whereas the RAI only has one. If we first find the exact supertype of
the ArrayImg WITH RESPECT to the RAI (thus getting RAI<FloatType>), we
can now easily infer the type of the I.