Continue when type var assignment agrees already
Often, when inferring type variable assignment, we come across a type
variable multiple times. When we do so, we only need to make sure that
what it should be assigned to in THIS instance is consistent with what
it has been assigned to in past instances. Once we have checked this, we
can simply continue on to the next type variable. The problem here was
that we were not doing so, which, with the help of the last commit,
     could cause some nasty StackOverflow exceptions.