Clean up size Ops
Before this commit, there existed the following size Ops:
1. Calculate size of IterableInterval
2. Calculate size of Iterable

If the user wished to compute the size of, for example, some RAI that
did NOT extend II, they would be stuck with the slow Iterable version of
the Op. To improve performance, we remove the II size Op, change the
name of the Iterable Op to IterableSize, and create a higher-priority
DefaultSize Op that operates on Intervals (which will cover both RAI and
II).

The DefaultSize Op is much faster than the IterableSize Op,
operating linearly in the number of dimensions of the Interval instead
of linearly in the number of pixels in the Image. Thus it is higher
priority.