Clean up Moment1AboutMean
This was a really convoluted way of setting the output to zero.

A simple proof that this is what the op was doing goes as follows. The
for loop within the Op was summing the differences between each value in
the iterable and the mean value of that iterable. This is equivalent
mathematically to summing all values within the iterable and then
subtracting n times the mean value of that iterable (where n is the
length of the interval).

But we know that by definition summing all values within the iterable
and then dividing by n gives the mean of the iterable. Thus, since we do
not divide by n after the for loop (this actually happens on line 70),
the summation of all items within the iterable will be n times the mean.
It thus follows that the difference between the sum of all values in the
iterable and n times the mean is zero. Thus res must be zero after the
for loop has completed, and the output must then be set to zero.

Instead of going through this logic, let's just save ourselves the
hassle and directly set the output to zero.