Multithread Moment2AboutMean
Sigh, if it were only so simple as to call stats.variance...

Actually, the only thing that is preventing us from doing so is the
division that happens at the end. The variance Ops divide by (size - 1),
while Moment2AboutMean divides by size.

At least we are able to multithread this computation.