Multi-thread moment Ops
Note that some of the expected values in ImageMomentsTest were changed.
This is due to the change in the math; we now perform the calculations
within Imglib2's type math, instead of in double math. This allows for
more precision, and allows us to use LoopBuilder. Many of the changes
were miniscule (i.e. changing the last two numbers of the expected
value), however since these numbers are large in magnitude the changes
were still greater than the allowed epsilon.