Change parameters of resolveOpDependencies()
We really only need the OpInfo and the Map<TypeVariable, Type> in order
to resolve dependencies.

Since we made it a requirement to pass an OpCandidate to this method, we
had to jump through some hoops when adapting. This saves us a bit of
work.

This commit also creates a method retaining the OpCandidate signature
for ease of use