Create simplification framework (#25)
* Create "converted" struct based on simplifiers

* Lean on matcher to find simplifed matches

The DefaultOpMatcher already has methods to return a single match from a
list of OpRefs, which means that if we change the return of simplifyOp
to a list of OpRefs, we can make use of this code.

* Isolate SimplifyTest

Now the tests depend on an Op within the test itself, which prevents
changes to external Ops from affecting the outcome of the test

* Convert Simplifiers to Plugins

This allows us to extensibly discover simplifiers

* Add simplifiers for all boxed primitive Numbers

* OpRef: only one op type allowed

OpRefs with more types are:
* More complicated
* Never used in practice

These things (not sure what they will be yet) calculate the magnitude of
the (worst-case) loss when converting between two types

* Clean DefaultOpEnvironment.simplifyInfo

* convert all conversionLoss behavior to ops

These Ops are now called LossReporters

* Improve Javassist simplified Op creation

This commit does two things. Firstly, it provides a proof of concept for
generic simplifiers. Secondly, it adds the logic necessary to simplify
to/from types that are not directly related to each other (as
illustrated in the test, where the simplification chain is Byte[] ->
ObjectArray<Number> -> Integer[])

* mature optimization

The startup time was greatly increased due to the simplification
process. We can mitigate a lot of this time by lazily initializing
elements of the SimplifiedOpInfos, such as their simplified type and
their simplified Struct.

* Only create simplified infos when necessary

We now create the simplifiedOpInfos for Ops whose name matches the Op
request. This lazy approach saves time and opens the possibility for us
to refactor Simplifiers into Ops

* WIP: Use Ops for simplification/focusing

* Remove optional parameter String in OpUtils

Optional parameters are no longer allowed

* Assume infinite loss when no lossReporters exist

* Test direct match to simplified OpInfo

* Improve retyping for edge case scenarios

In rare edge cases, functional interfaces don't declare their abstract
method. This might happen for a convenience interface like:

interface AllSameFunction<T> extends BiFunction<T, T, T> {}

such an interface does not have a declared abstract method, so we have
to find it via its superinterface BiFunction. The method we get out of
BiFunction, however, will not know of the mapping of the Ts in
AllSameFunction to the T, U, R of BiFunction; this is a bug in the
method.

This commit adds logic that determines the mappings of the type
variables of interfaces like AllSameFunction to the type variables of
BiFunction, so that Ops implementing AllSameFunction can be retyped.

* Remove obsolete data structure

This data strucutre held the Simplifiers, which we no longer use (since
    we converted to an Ops-based simplification framework)

* Clean DefaultOpEnvironment

* Purge commented code

* Expose problem with non-reifiable lambdas

* Add simplification to mutate input arguments

* Clean code

* Extract simplified Op creation to utility class

* Remove unused import

* Add synthetic check to reification

* Add example tests for inplace simplification

* Take the work out of SimplifiedOpCandidate

SimplifiedOpCandidate was taking on too much work. I would rather this
work be extracted to a helper class, so that the SimplifiedOpInfo can be
called without needing the SimplifiedOpCandidate. Of course,
SimplifiedOpCandidate can't go away, since some functionality (like
determining loss) is probably better performed there.

* Test simplification priority

* Improve determination of suitable Op wrapper

When Op wrappers exist for both a class and one of its superclasses, Op
wrapping would fail as multiple wrappers would be suitable. The Op
wrapping logic was not smart enough to filter these matches in any way.

This commit adds a method to filter superclasses out of the list of
suitable wrappers. Since these wrappers also preserve the exact
supertype of the Op w.r.t. the wrapper class, it is desirable to have
subclasses declare their own wrappers, and this fix prevents multiple
suitable wrappers from preventing a wrapper match (so long as all
classes are related).

This commit also fixes a bug in the error message when multiple wrappers
would match. The error was designed to print the classes of each wrapper
that matched, but would instead vomit out the classes of all known
wrappers. This bug has also been fixed.