Introduce separate @Container for computer outputs
They aren't exactly mutable in the same was as the mutable inplace
arguments are. Rather, they are preallocated empty containers whose
contents will be filled in by the invoked computation.

My strong intuition is that having this distinction baked into the type
annotations, and accessible at runtime, may be somehow helpful later.

This also adds a corresponding ItemIO.CONTAINER value for struct
members; the ItemIO.BOTH value is retired in favor of ItemIO.MUTABLE,
which corresponds to the @Mutable annotation for the Inplace interfaces.