Make performance improvements
This commmit makes two performance improvements. Firstly, it removes the
need to store the inputs. What is the point of storing the inputs, if
you know that the inputs are never modified? If you want to know how a
particular Object came to be the way it is, just look at the Ops that
modified the Object i.e. the Ops for which this Object was an output

Secondly, change the backing Object of the OpHistory to a LinkedDeque.
This allows us to add faster to the end, removing the need to iterate
through the history in order to add something new