refactor findOpInstance
There are benefits to reworking the caching system, and to refactor the
functionality of findOpInstance. The goal is to have findOpInstance
place Ops DIRECTLY into the cache, and NOT return anything. Then Ops can
be returned directly from the cache. This gives us more control over Op
wrapping and retains a more "raw" version of the Op in the cache. We
then push Op wrapping to the cache.

TODO: Use the hints to only record executions of top-level Ops. What
works is to be able to re-wrap an Op with a new UUID. What we DON'T yet
do is re-wrap its dependencies with the same UUID.