Extract OpRef interface
This commit also makes some changes to SimplifiedOpRef that make it more
correct. I chose to throw UnsupportedOperationException for many of
the standard OpRef methods because the concrete type that the
SimplifiedOpRef is asking for is rather indeterminate - it is really
asking for an Op that is of a set of Types, so to pin a particualar set
of arguments, a particular output type, etc. is not possible until you
have a SimplifiedOpInfo to narrow down these Types. Similarly, we throw
an UnsupportedOperationException for typesMatch because the typing of
the SimplifiedOpRef is unclear. We could match against all possible
permutations of the input/output types, but we don't really have enough
information about how to form the opType. We could probably find a way
to devise all opType permutations, but this would be a pretty
computationally expensive method to run, so for now we'll just throw an
error