Absorb net.imagej.ops2.geom.geom3d.mesh
There seemed to be no reason for this package. The contents did not
really deal with Meshes, and many Ops in the net.imagej.ops2.geom.geom3d
package (that actually imported Mesh) depended on this subpackage (which
is bad). For these reasons, let's just absorb the mesh subpackage into
the geom3d package