More signature->InfoChain work
The most important change here is the change to the javassist portion of
OpMethodInfo. The current framework cannot cache the Op returned if
called via signature, as we do not have enough to cache a
MatchingConditions, but this commit avoids the matching if we have
alreaady created the class. This commit should be viewed with scrutiny.