Retain RichOp when resolving dependencies
This allows the dependencies to update the progress of their parent Task

This commit also fixes some expected answers within ProvenanceTest.
Previously, these tests asserted that only the outside (mapper) function
altered the output. But, in this case, the output was directly created
by the mapped function. Thus there were really two Ops that directly
altered the output. Now, the test reflects that.