Flesh out the README
This is a draft of documents I've been sitting on for quite a while.
I'm making them public now so that people who come across this
repository have any clue whatsoever what we are trying to do here.
But the documents are by no means complete, or fully coherent yet.