Clean up OpUtils
This commit began as a migration of OpUtils back to SciJava Ops Engine.
But since writing this patch, that class already moved back to
scijava-ops-engine in fd6f5101307880719577e05a2f82a98803062bc6. So this
patch is now just remnants of that work, in case any of it is useful.

Signed-off-by: Curtis Rueden <ctrueden@wisc.edu>
Signed-off-by: Mark Hiner <hinerm@gmail.com>