Purge imagej-common dependency
We need to purge imagej-common as a prerequisite to purging
scijava-common usage, since imagej-common depends on scijava-common.

We were barely using ImgPlus, and are not using Dataset.

We *are* using BigComplex and UnboundedIntegerType though,
so those have migrated into imagej-ops2 from imagej-common.
Ideally, these classes would become part of the imglib2 library layer
somewhere, so their presence in imagej-ops2 might be temporary.

Co-authored-by: Curtis Rueden <ctrueden@wisc.edu>