OpBuilder: separate out arity choice
Instead of having a single choice for input/inType for all 16 arities (a
total of 49 methods), we make the first choice arity 0-16 and then
narrow to the 3 input/inType choices.

Although this makes more method calls in the builder chain, the hope is
to avoid overwhelming developers with large numbers of similar method
calls.