OpBuilder: fix newlines in conditional javadoc
If the full line is conditional, the #end should appear on the following
line (so the conditional line includes a newline character).

If only the final part of the line is conditional, so that the #end
appears at the end of the line, a standalone line-break must appear
after the #end. This looks unintuitive in the template, but is the only
way I could find to avoid Velocity eating the newline otherwise.

Inline conditionals that do not appear at the end of the line are
unaffected, as the newline character is not enclosed in the if..else.