OpEnvironment: add arity convenience methods
Specifying the "arity1, 2, 3, etc..." part of the builder is a pain if
you are more familiar with ops. By adding convenience methods with
common "arity" names (e.g. "unary", "binary", etc..) bypasses the need
to make this extra call within the builder process.

Note that this is not perfect, as changing the number of arguments will
still necesitate a change to the "arity" portion of the builder call.
But this is a compromise.

Note that the moving of "Hints" to OpEnvironment coupled with this
change now means there are 32 new signatures.