Refactor all RAI lifters
Instead of operating on Imgs, all simple and complex lifters are
rewritten to be genericized operating on <T extends RandomAccssible
Interval> which will allow lifting to operate on any RAI subclass.

This commit also consolidates RAI ops (RAItoIIOps) to the adapt package
instead of types/adapt.

Lifting has further been generalized to operate on functions not just
with all RAI params, but combinations of RAIs and Types.