RSMR: add defensive exception guard
IllegalStateException can come back from isAssignable when there's an
unmatched type param in the base type (at least).

We don't need to stop the op search in this case, just move on from that
particular candidate.

Other solutions would be to return false instead of throwing that
exception, in Types.isAssignable. We could also catch any exception and
not just IllegalState. But this seemed like the most minimal change that
would not affect any other code path that might rely on exceptions being
thrown.