OpBuilder: default to Any return value
When the OpBuilder doesn't have an output/outType specified, we were
defaulting to Object. However this can incorrectly lead to matching
scenarios where we look for Ops with return values that Object can
be *cast to* (which will almost universally fail).

Instead we want to match Any, in that any return value is acceptable.

This change requires updates to Types/GenericAssignability to support
such Any return value scenarios, as they were checking "instanceof Any"
but not if it was Any.class (which is the scenario of the OpBuilder
using Nil<Any>, instead of new Any(), which was determined to be
undesirable).