Ensure simplified, reduced Op Infos are collapsed
There already exists logic in the SimplifiedOpDescriptionGenerator to
collapse reduced Op infos (i.e. Ops with nullable parameters), such that
only one instance of reduced Op infos exist in the help calls.

This collapsing logic did not apply to reduced Op infos that were
simplified, and in this case we'd have reduced Op infos show up many
times. This commit fixes that, and adds a test that fails without the
changes.