WIP: Attempt to JPMSify the scijava-ops-opencv tests
It does not quite work as written.

Here are some notes I wrote as I was developing this patch:

Adding a scijava-types dependency to main causes a weird problem where
the unit tests then fail to execute, due to a sudden inability to locate
classes from org.scijava.ops.api.

To fix the issue, I made a separate module-info.java for the tests, to
require the org.scijava.ops.api because the tests *do* need it.

Aside: one thing that's weird is we weren't depending on the jupiter
engine. So I just tried adding that. But the failures are the same.

The maven-surefire-plugin documentation has a page on running unit tests
for JPMS-enabled projects:
https://maven.apache.org/surefire/maven-surefire-plugin/examples/jpms.html

The examples given there use a distinct module name for the test
module-info.java, and also require the main module. So I tried changing
the test module-info.java of this patch to be structured this way, with
its own module name (I appended .test), and moved the TestOpenCV class
into a test subpackage accordingly to avoid package clash across
multiple modules... but I get the same failures afterward. So I changed
it back to the same module declaration as the main module-info.java,
since that seems less broken.

I think the issue with kotlin is that scijava-io-http depends on okhttp
which depends on okio which depends on kotlin stuff (because okio is a
KMP project now), and scijava-io-http is an unnamed module without a
module-info.java. So we are stuck guessing how the module config should
look for that piece of our dependency tree, I suppose.

My guess is that when there is no module-info.java in the src/test, it
runs the tests outside JPMS, i.e. with a plain classpath. Whereas once
you have a test-side module-info, the surefire plugin thinks "ah, you
want to test this as a JPMS module, I should use a module path."

Which suggests to me that we are going to be in big trouble as soon as
anything downstream tries to use e.g. scijava-ops-opencv as a JPMS
module...

The fix might be a simple as just adding all the test scope dependencies
to the test module-info.java, as I should have but didn't.

I was able to get much closer to a working JPMS configuration for
scijava-ops-opencv, but the openblas native library dependency is not
something I know how to overcome.

The other crazy thing I had to do was exclude okio and explicitly add
only okio-jvm. Otherwise, JPMS gets confused about which JAR has the
okio module, since they both purport to have it (but really, only
okio-jvm has the needed classes).

I think the lesson here is: JPMS projects can't fully depend on
non-JPMSified projects that do unusual things, such as wrap native
libraries into JAR files and extract them. Or even non-JPMS-friendly
things, like having multiple JAR files with the same package prefix
and/or module name. I say "fully" depend because yeah, you can depend on
them, but if they don't work in a modular context, you'll be stuck using
your code only in a non-modular (i.e. classpath) style.

This is probably worth a forum post in Development category on the
forum, and a heads up to Tobias that JPMSifying ImgLib2 is going to
cause these headaches for any projects that have third party
dependencies. Getting things like BigWarp and SNT into JPMS is not going
to be feasible any time soon.

I noticed that okio 3.8.0 does not have this double-module-declaration
issue anymore. But upgrading to 3.8.0 still does not avoid the
`NoClassDefFound okio/Buffer`, for reasons I don't understand. I thought
it was happening because okio.jar had a clashing `Automatic-Module-Info`
line with okio-jvm, but nope, apparently that's not why. Excluding the
former JAR does fix the error though, so I dunno. I guess we could
exclude it in scijava-io-http and release a new version of that, so that
JPMS projects built on scijava-ops can use scijava-io-http more easily.

I found an SO post about exactly this JavaCPP + JPMS problem:
https://stackoverflow.com/a/71385521/1207769. However, I think the
solution detailed there boils down to passing `--add-modules
ALL-MODULE-PATH` to the test execution (if I'm reading it correctly),
which seems like not a proper solution.

Giving up for now.