Move Op suggestion generation to top-level call
Op matches generate a large history of exceptions, as there are multiple
matchers and only one will ultimately succeed (if any). Generating
potential alternatives for the user is a helpful way to point them to
ways to correct their Op call, but doing so for each matching attempt
means we then need to display all the failures to the user, to the point
of being unhelpful. It also means an order of magnitude more Op matches
are attempted for any match request that is destined to fail, as each
failing request goes through a help call to identify near matches.

Instead, we now make this help call once at the complete end of
matching. The agglomerated exceptions from matching are logged at a
debug level but then discarded, leaving a single simple stack trace with
the resulting help text (presumably to display to users).

This may result in a loss of information, as there are no help requests
for any converted parameters or adaptation attempts. If that is
determined to be useful and necessary then it could be built into the
help functionality.

Closes https://github.com/scijava/scijava/issues/184