Further simplify error messages
This change completely removes the individual matcher failures from the
top-level error reporting, instead standardizing to just report the
failed request.

To avoid a total loss of information we special-case some failures
(duplicate Ops).

Note that this does require generalizing some of the exception tests
(although this could be expanded if there was an easy way to capture
debug log output in the tests).

Also removes empty lines between Op descriptions (when present) and
parameters.