Do not modularize the scijava-taglets component
Due to a bug in the maven-javadoc-plugin, the following warning gets
emitted with the scijava-taglets artifact included in the javadoc build:

  [WARNING] Unable to auto-detect Taglet class names from
  '/path/to/jar.jar'. Try to specify them with <taglets/>.

Running with `mvn -X` shows the underlying cause:

  NoClassDefFoundError: module-info is not a
  class because access_flag ACC_MODULE is set

Removing scijava-taglets's module-info.java avoids the problem.

I filed an issue about it on the Apache JIRA:
https://issues.apache.org/jira/projects/MJAVADOC/issues/MJAVADOC-794

It's too bad, because ideally we would leave the module-info.java in
place and simply not export anything from the scijava-taglets module
to completely encapsulate it, since the taglets do still function
correctly with the <taglet> elements manually specified in the POM.
Hopefully, a future version of the plugin will fix this problem,
at which point we can modularize the scijava-taglets component again.