Move Types.isApplicable* methods to MatchingUtils
The isApplicable methods are *almost* not used at all anymore...
only the inputsMatch method of RuntimeSafeMatchingRoutine calls it as
part of a final check to evaluate input suitability of op candidates,
as part of its typesMatch(OpCandidate) method.

But according to @gselzer, it's actually the *other* typesMatch method,
typesMatch(Type, Type, Map<TypeVariable<?>, Type>), which is the real
workhorse of the system, performing the primary type matching check.

In lieu of fully cleaning up this mess, this commit moves the
almost-obsolete methods from the Types class, where they were public
and exported, to the MatchingUtils class, where they are internal and
encapsulated. Later, we can tweak the internal scijava-ops-engine code
to fully eliminate these methods. But in the meantime, at least they
won't become permanent public API of the scijava-types module.