Remove Progress reporting from op discovery
Due to an as-yet ill-understood bug, this logic results in the
OpRegressionTest to fail on one of my Linux systems with:

  java.lang.ExceptionInInitializerError: Exception
  java.lang.IllegalStateException: Task OpEnvironment:
  Discovering Ops has subtasks that did not complete!

@gselzer and I thought this error was simply a consequence
of a previous Op parsing error thrown by snakeyaml:

  Caused by: java.lang.ExceptionInInitializerError: Exception
  org.yaml.snakeyaml.scanner.ScannerException: while scanning for the
  next token; found character '@' that cannot start any token.
  (Do not use @ for indentation);  in 'reader', line 26008, column 2:
    {@code <C>}
     ^ [in thread "main"]
    at org.yaml.snakeyaml@2.0/org.yaml.snakeyaml.scanner.ScannerImpl.fetchMoreTokens(ScannerImpl.java:445)
    ...
    at org.yaml.snakeyaml@2.0/org.yaml.snakeyaml.Yaml.load(Yaml.java:434)
    at org.scijava.ops.engine@0-SNAPSHOT/org.scijava.ops.engine.yaml.impl.YAMLOpInfoDiscoverer.parse(YAMLOpInfoDiscoverer.java:104)
    at org.scijava.ops.engine@0-SNAPSHOT/org.scijava.ops.engine.yaml.impl.YAMLOpInfoDiscoverer.lambda$discover$0(YAMLOpInfoDiscoverer.java:78)

which we found was caused by the scijava-ops-image module's
generated target/test-classes/op.yaml file being somehow mangled,
with part of at least one op declaration block missing, and instead
the sequence `{@code <C>}` alone on a line with no indentation.
In other words: indeed, the YAML was invalid.

However, upon removing the Progress reporting logic with this commit,
the mangled op.yaml generation has not happened on the affected system
again, whereas previously it was reliably failing.

More investigation is needed to better understand what's going on.
It may be that when the YAML fails to parse, it cascades into the
DefaultOpEnvironment discovery, resulting in the Progress.complete()
method never getting called, and thus that task remaining stuck in the
map data structure as a zombie, and then when then next
DefaultOpEnvironment construction happens, its Progress.complete() also
fails because it erroneously believes it has subtasks remaining, when in
actuality the "subtasks" are really just previously aborted
DefaultOpEnvironment construction attempts. Then again, it may be that
something else is going on; if the narrative here is correct, we would
expect to see the Progress.complete() call fail with *1* pending
subtask, but it fails with *3* pending subtasks...