scijava-ops-image: rewrite open module comment
It's not a hack, exactly. Individually opening every one of the 105
org.scijava.ops.image.** packages is very verbose, and also does not
fully work on its own to let all tests pass. Maybe there are also
scijava.surefire.args flags that need to be set? But rather than mess
with that, we stick with the solution that works via a single keyword.