Add improved SciJava Ops Python gateway
This commit adds an improved Python gateway contained
in the ops-gateway.py file. This gateway resolves a
usability bug where nested namespaces like
"features.haralick.asm" are unreachable as the string
"haralick.asm" is appended as an attribute to the gateway instead
of the "haralick" namespace being added as an intermediate namespace.
Attempting to access "haralick" fails as that attribute does not exist.
Additionally this commit adds the scijava-ops-flim library to the
gateway.

Intended use cases/scenarios
  - A user can simply run `python -i ops-gateway.py` to obtain an
    actiave `ops` environment.
  - A user can copy the ops-gateway.py module into their own project and
    create an ops gateway for their own internal project use.