POM: revise comments about component versions
They are not temporary. Even once release versions are added to
pom-scijava, the SNAPSHOT overrides here will need to remain; the
purpose of pom-scijava is to offer a stable bill of materials, whereas
the purpose of the declarations here is to link the modules for
Maven's multi-module build. The former requires release versions only,
while the latter requires snapshot versions for non-release commits.